separated js code into modules and changed styling of interface
