+++ index.html
@@ -26,12 +26,13 @@
 
     <div id="instructions">
       <ul>
-        <li>spacebar: start/stop</li>
-        <li>f: toggle fullscreen</li>
+        <li>SPACEBAR - start/stop</li>
+        <li>F - toggle fullscreen</li>
+        <li>S - set sprint intervals</li>
       </ul>
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="index.js"></script>
 
   </body>
 </html>
